blind people cant see sign language (#6614)
