+++ Content.Goobstation.Common/Chat/ChatMessageOverrideInRangeEvent.cs
@@ -1,0 +1,10 @@
+namespace Content.Goobstation.Common.Chat;
+
+[ByRefEvent]
+public record struct ChatMessageOverrideInRange(bool RequiresHearing = false, bool RequiresSight = false, bool Cancelled = false)
+{
+    public void Cancel()
+    {
+        Cancelled = true;
+    }
+}
