--- Content.Goobstation.Common/Chat/ChatMessageOverrideInVoiceRangeEvent.cs
@@ -1,10 +1,0 @@
-namespace Content.Goobstation.Common.Chat;
-
-[ByRefEvent]
-public record struct ChatMessageOverrideInVoiceRange(bool Cancelled = false)
-{
-    public void Cancel()
-    {
-        Cancelled = true;
-    }
-}
