+++ Content.Shared/_EinsteinEngines/Language/LanguagePrototype.Goob.cs
@@ -1,0 +1,10 @@
+namespace Content.Shared._EinsteinEngines.Language;
+
+public sealed partial class SpeechOverrideInfo
+{
+    /// <summary>
+    ///     If false, an entity can hear this language even when it's unable to see (i.e. blind),
+    /// </summary>
+    [DataField]
+    public bool RequireSight = true;
+}
