+++ Resources/Prototypes/_EinsteinEngines/Language/genericlanguages.yml
@@ -305,6 +305,7 @@
   speech:
     allowRadio: false
     requireSpeech: false
+    requireSight: true # Goob - the blind dont see
     requireLOS: true # Floofstation - Check Line-Of-Sight
     color: "#dddddd"
     messageWrapOverrides:
